Geometrical boundaries stored as double
Fixes a problematic situation in which a Particle can find itself stuck
between domain borders if both its position and the boundaries are
stored as floats. In future commits, one of these quantities may
reverted to float for performance reasons.

+++ include/nodes.hpp
@@ -41,7 +41,7 @@ class Node {
         SError InitAllLeafs();
         std::array<double,2> Center() const;
         std::array<double,4> GetBounds() const;
-        SError SetBounds(float x1, float x2, float y1, float y2);
+        SError SetBounds(double x1, double x2, double y1, double y2);
         virtual SError Decompose( std::vector<Particle*>& parts );
         virtual SError TimeEvolution( double dt );
         unsigned int GetLevel() const;
